+++ packages/priority-queue/src/__tests__/heappushpop.test.ts
@@ -1,4 +1,5 @@
-import { Comparator, heappushpop } from '..'
+import { heappushpop } from '..'
+import { Comparator } from '@structured/comparable'
 import '../../setupTests'
 
 describe('heappushpop', () => {